+++ History.md
@@ -2,6 +2,12 @@
 
 ## vNext
 
+## v0.0.11
+
+* Made .get() reactive on both the client (e.g. for routing) and server
+(e.g. for publish), and added a ready() method for the same reasons.
+* Server: support for publish functions
+
 ## v0.0.10
 
 * Added `headers.ready()` function to provide a callback to be run once we have